+++ packages/uikit/src/components/FAuthMethodModal/FAuthLinks.vue
@@ -1,25 +1,32 @@
 <template>
-  <div class="f-auth-mixin" :class="{ 'f-auth-mixin--small': smAndDown }">
-    <div class="f-auth-mixin__left">
+  <div
+    class="f-auth-links f-auth-step2"
+    :class="{ 'f-auth-step2--small': smAndDown }"
+  >
+    <div class="f-auth-step2__left">
       <template v-if="qrUrl">
-        <f-qr-code class="f-auth-mixin__qr" :text="qrUrl" :size="182" />
+        <f-qr-code class="f-auth-step2__qr" :text="qrUrl" :size="182" />
       </template>
-      <f-loading v-else :fullscreen="false" class="d-flex" />
+      <f-loading
+        v-else
+        :fullscreen="false"
+        class="d-flex f-auth-step2__loading"
+      />
     </div>
-    <div class="f-auth-mixin__right">
-      <div class="f-auth-mixin__title">
+    <div class="f-auth-step2__right">
+      <div class="f-auth-step2__title">
         {{ labels[0] }}
       </div>
-      <div class="f-auth-mixin__subtitle">
+      <div class="f-auth-step2__subtitle">
         <span v-html="labels[1]"></span>
       </div>
       <f-button
         color="greyscale_1"
-        class="f-auth-mixin__icon"
+        class="f-auth-step2__icon"
         @click="handleInstall"
       >
         <v-icon size="16" color="greyscale_7" class="mr-1"> $install </v-icon>
-        <span class="f-auth-mixin__install_label greyscale_7--text">
+        <span class="f-auth-step2__install_label greyscale_7--text">
           {{ labels[2] }}
         </span>
       </f-button>
@@ -69,7 +76,7 @@ class FAuthLinks extends Vue {
       $t(
         this,
         "links_connect_detail",
-        "<a class='f-auth-mixin__link' href='https://getlinks.jp/'>Links",
+        "<a class='f-auth-step2__link' href='https://getlinks.jp/'>Links",
         "<svg width='14' height='14' viewBox='0 0 14 14' fill='none' xmlns='http://www.w3.org/2000/svg'>" +
           "<path d='M12.2068 1.79276C10.2975 -0.116585 7.2018 -0.116585 5.29246 1.79276L4.64288 2.44234C4.25236 2.83286 4.25236 3.46603 4.64288 3.85655C5.03341 4.24707 5.66657 4.24707 6.0571 3.85655L6.70667 3.20698C7.83497 2.07868 9.6643 2.07868 10.7926 3.20698C11.9209 4.33528 11.9209 6.16461 10.7926 7.29291L10.1425 7.94304C9.75195 8.33357 9.75194 8.96673 10.1425 9.35726C10.533 9.74778 11.1662 9.74778 11.5567 9.35726L12.2068 8.70712C14.1162 6.79778 14.1162 3.70211 12.2068 1.79276Z' fill='#076aff'/>" +
           "<path d='M3.85726 6.05638C4.24779 5.66586 4.24779 5.03269 3.85726 4.64217C3.46674 4.25165 2.83357 4.25165 2.44305 4.64217L1.79236 5.29286C-0.116985 7.2022 -0.116985 10.2979 1.79236 12.2072C3.70171 14.1166 6.79738 14.1166 8.70672 12.2072L9.35713 11.5568C9.74765 11.1663 9.74765 10.5331 9.35713 10.1426C8.9666 9.75207 8.33344 9.75207 7.94291 10.1426L7.29251 10.793C6.16421 11.9213 4.33488 11.9213 3.20658 10.793C2.07828 9.6647 2.07828 7.83537 3.20658 6.70707L3.85726 6.05638Z' fill='#076aff'/>" +