+++ packages/uikit/src/services/mvm/abis/index.ts
@@ -1,5 +1,0 @@
-import AssetABI from "./asset.json";
-import BridgeABI from "./bridge.json";
-import RegistryABI from "./registry.json";
-
-export { AssetABI, BridgeABI, RegistryABI };