+++ packages/uikit/types/utils/payment.d.ts
@@ -2,23 +2,20 @@ import type { VueConstructor } from "vue/types/umd";
 import type Vuetify from "vuetify/lib";
 
 export interface PaymentOptions {
-  url?: string;
-  code?: string;
-  multisig?: boolean;
-  fennec?: Fennec;
-  data?: {
-    recipient: string;
-    assetId: string;
-    amount: string;
-    traceId: string;
-    memo: string;
+  scheme: string;
+  channel: "mixin" | "fennec" | "metamask" | "walletconnect";
+  hideCheckingModal?: boolean;
+  actions: {
+    mixin: () => Promise<boolean>;
+    fennec: () => Promise<boolean>;
+    mvm: () => Promise<boolean>;
   };
-  info?: {
+  info: {
     symbol: string;
     logo: string;
     amount: string;
   };
-  checker: () => Promise<any>;
+  checker: () => Promise<boolean>;
 }
 declare function Payment(): void;
 declare namespace Payment {